+++ accessories/BLEAIOS/aiosAPI.py
@@ -6,26 +6,24 @@ import time
 import asyncio
 from bluepy import btle
 
-if __name__ == "__main__":
+if __name__ == "__main__":  # testing this module
 	import sys,os,logging
-	sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), '../..'))
+	sys.path.append(os.getcwd()) # bring lib in path: to be called from cwd=fshome
+	#sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), '../..')) # bring lib in path
 	import lib.tls as tls
 	logger = tls.get_logger(__file__, logging.DEBUG)
 else:
 	import lib.tls as tls
 	logger = tls.get_logger()
-	#import logging
-	#logger = logging.getLogger()
 from accessories.BLEAIOS.bluepyBase import bluepyDelegate,BAS_SVR,showChars
 
 DEVADDRESS = "d8:59:5b:cd:11:0c"    # to be adapted to address of your device
-AIOS_SVR = "00001815-0000-1000-8000-00805f9b34fb"  # automation-IO service
+AIOS_SVR = 0x1815  # "00001815-0000-1000-8000-00805f9b34fb"  # automation-IO service
 ENV_SVR  = "6c2fe8e1-2498-420e-bab4-81823e7b0c03"  # environmental service as defined in BLE_automation
 
 mdOUT = 0b00	# following GATT AIOS for dig bit modes
 mdINP = 0b10
 mdNOP = 0b11
-
 
 chTEMP=1
 chHUMI=2
@@ -54,21 +52,39 @@ NAMES ={chTEMP:'temperature', chHUMI:'humidity', chECO2:'CO2', chTVOC:'VOC',
 		chDIGI:'digitalIO', chBAT:'devBatteryLev', chANA1ST:'AnalogChan'}
 
 class aiosDelegate(bluepyDelegate):
+	''' specialization of bluepy to have aios interface '''
 	def __init__(self, devAddress=DEVADDRESS, scales=SCALES, chInpBits={chDIGI:16}, loop=None):
+		''' devAddress : address of device as found by e.g. bluetoothctl
+		    scales : scale factor for a quantity to have Si units
+		    chInpBits : defining active input bits for notification
+		    loop : asynchronous loop used by async methods defined here
+		'''
 		super().__init__(devAddress, scales, loop)
 		self.chInpBits=chInpBits
 		self.digmods=[]
 		self.bitvals=[]
 		self.anamap={} #self._getAnaMap()
+		self.digPulses={}
+		self._digiParse(self._readDigitals())
+		if loop:
+			loop.create_task(self.pulseHandler())
+		else:
+			asyncio.create_task(self.pulseHandler())
 		
 		if chInpBits:
 			for chB,bitn in chInpBits.items():
 				logger.info('set inp on %d for chId %d' % (bitn,chB))
 				self.setDigMode(bitn, mdINP, False)
 		self._sendDigBits()
-		logger.info('dev %s state:%s' % (self.dev, self.dev.getState() if self.dev else None))
+		logger.info('BLEdev %s state:%s' % (devAddress, self.dev.getState() if self.dev else None))
 
+	def __repr__(self):
+		"""Return the representation of the ble client."""
+		return 'ble aios client: dev@{} digitals={} analogs={}>' \
+			.format(self.dev.addr if self.dev else "nc", self.digmods, self.anamap)
+	
 	def _CharId(self, charist, CharDef=CHARS):
+		''' return unique id for a chracteristic '''
 		chId = None
 		uuid = charist.uuid
 		self._getAnaMap()
@@ -84,9 +100,11 @@ class aiosDelegate(bluepyDelegate):
 		return chId
 
 	def startChIdNotifyer(self, chId):
+		''' start GATT notification mode on server for denoted characteristic '''
 		if self.dev is None:
 			logger.error('no ble device for notifying on %d' % chId)
 			return
+		logger.info('starting notification on %s' % chId)
 		if chId>=chANA1ST:  # finding which chan
 			hand = self._getAnaMap(chId - chANA1ST)
 			if hand:
@@ -102,10 +120,14 @@ class aiosDelegate(bluepyDelegate):
 		elif chId == chBAT:
 			service = self.dev.getServiceByUUID(btle.UUID(BAS_SVR))
 		if chId in CHARS:
-			charist = service.getCharacteristics(btle.UUID(CHARS[chId]))[0]
-			self.startNotification(charist)	
+			charist = service.getCharacteristics(CHARS[chId])
+			if charist:
+				self.startNotification(charist[0])
+			else:
+				logger.error("no BLE characteristic for %s with uuid:%s in %s" % (chId,btle.UUID(CHARS[chId]),service))
 				
 	def _getAnaMap(self, anaChan=None):
+		''' get / create map between analog channel number and characteristic handle '''
 		if not self.anamap:
 			hand=9999
 			descr=None
@@ -114,6 +136,7 @@ class aiosDelegate(bluepyDelegate):
 				descr = self.dev.getDescriptors()  # also having the characteristics
 			if not descr:
 				logger.warning("no descriptors for dev %s" % self.dev)
+				return None
 			for des in descr:
 				if des.uuid == btle.UUID(CHARS[chANA1ST]):  # it is an analog channel charist
 					hand = des.handle
@@ -130,30 +153,28 @@ class aiosDelegate(bluepyDelegate):
 			return self.anamap[anaChan]
 		return None
 		
-	def _getAnaMap0(self):
-		descr = self.dev.getDescriptors()  # also having the characteristics
-		anamap={}
-		for des in descr:
-			if des.uuid == btle.UUID(CHARS[chANA1ST]):  # it is an analog channel charist
-				hand = des.handle
-				charist=self.dev.getCharacteristics(hand-1,hand)[0]
-				presform = charist.getDescriptors(btle.UUID(CHARS[dscPRESFORM])) # takes some time
-				logger.debug('(%d) ana prfrm:%s with presfrm:%s ' % (hand, presform[0].handle, presform))
-				if presform:
-					datPresForm = presform[0].read()
-					chan == datPresForm[5]
-					anamap[chan] = hand
-		return anamap
 
 	def _getAnaCharacteristic(self, chan):
+		''' get analog characteristic which is mapped to denoted channel '''
 		hand = self._getAnaMap(chan)
 		if not hand:
 			logger.warning("analog chan %d not in map %s" % (chan,self.anamap))
 			return None
 		charist=self.dev.getCharacteristics(hand-1,hand)[0]
 		return charist
+		
+	def _readDigitals(self):
+		if self.dev:
+			service = self.dev.getServiceByUUID(btle.UUID(AIOS_SVR))
+			chT = service.getCharacteristics(btle.UUID(CHARS[chDIGI]))
+			if chT:
+				bts = self.read(chT[0])
+				logger.debug('reading digitals:%s' % tls.bytes_to_hex(bts))
+				return bts
+		return None
 
 	def setAnaVoltRange(self, chan, volt, SCL=SCALES[chANA1ST]):
+		''' set the reference voltage for the analog channel on the aios device '''
 		charist = self._getAnaCharacteristic(chan)
 		maxV = None
 		if charist:
@@ -165,12 +186,11 @@ class aiosDelegate(bluepyDelegate):
 				logger.info('dscVALRNG : %s min=%f max=%f' % (minmax, minV,maxV))
 				if volt and descr:
 					minmax = (int(minV*SCL) ) + (int(volt*SCL)  << 16)
-					minmax = (int(chan) ) + (int(volt*SCL)  << 16)	# hack to know channel at server
+					#minmax = (int(chan) ) + (int(volt*SCL)  << 16)	# hack to know channel at server
 					minmax = minmax.to_bytes(4, 'little') 
 					logger.info("update volt range chan:%d to %fV :minmax=%s on %s" % (chan,volt,tls.bytes_to_hex(minmax),descr[0].uuid))
 					self.write(descr[0], minmax)
 		return maxV
-
 
 	def _extBitsLen(self, nbits):
 		''' extend size of bit storage '''
@@ -178,12 +198,16 @@ class aiosDelegate(bluepyDelegate):
 			self.digmods.extend([mdNOP] * (nbits-len(self.digmods)))
 		if len(self.bitvals) < nbits:
 			self.bitvals.extend([False] * (nbits-len(self.bitvals)))	
-				
+
 	def _digiParse(self,digbits):
-		digbits =bytearray(digbits)
-		nbits = len(digbits)*4  # 4 bits per byte
-		bno =[]
-		self._extBitsLen(nbits)
+		''' parse received digital GATT structure '''
+		bno=[]
+		if digbits:
+			digbits =bytearray(digbits)
+			nbits = len(digbits)*4  # 4 bits per byte
+			self._extBitsLen(nbits)
+		else:
+			nbits=0
 		for bti in range(nbits):
 			bit2 = digbits[bti >> 2] >> ((bti & 3)*2)
 			if (bit2 & 2) == 0:
@@ -193,15 +217,17 @@ class aiosDelegate(bluepyDelegate):
 				self.bitvals[bti] = True if bit2 & 1 else False
 				if bit2 & 1:
 					bno.append(bti)
-		logger.info('digi parse:%s inp1:%s'  % ('.'.join('{:02x}'.format(x) for x in digbits),bno))
+		if nbits:
+			logger.info('digi parse:%s inp1:%s'  % ('.'.join('{:02x}'.format(x) for x in digbits),bno))
 		return bno
-		
+
 	def getDigBit(self,bitnr):
 		''' bits stored in little endian order i.e. low bits first '''
 		self._extBitsLen(bitnr+1)
 		return  self.bitvals[bitnr] 
 		
 	def _sendDigBits(self):
+		''' send state of digital bits to aios device '''
 		nbits = len(self.digmods)
 		if nbits<=0:
 			return
@@ -215,8 +241,10 @@ class aiosDelegate(bluepyDelegate):
 			service = self.dev.getServiceByUUID(btle.UUID(AIOS_SVR))
 			chT = service.getCharacteristics(btle.UUID(CHARS[chDIGI]))
 			self.write(chT[0], bytes(bitsbuf))
+			logger.debug('aios sending %d digdat:%s' % (nbits,tls.bytes_to_hex(bitsbuf,'.')))
 
 	def setDigBit(self, bitnr, val, updateRemote=True):
+		''' set state of a digital bit '''
 		self._extBitsLen(bitnr+1)
 		if self.digmods[bitnr] == mdNOP:
 			self.setDigMode(bitnr, mdOUT, False)
@@ -225,11 +253,46 @@ class aiosDelegate(bluepyDelegate):
 			self._sendDigBits()
 
 	def setDigMode(self, bitnr, mode, updateRemote=True):
+		''' set mode of a digital bit 
+		    mode = one of mdXXX constants
+		'''
 		self._extBitsLen(bitnr+1)
 		if self.digmods[bitnr] != mode:
 			self.digmods[bitnr] = mode
 		if updateRemote:
 			self._sendDigBits()
+			
+	def setDigPulse(self, bitnr, duration):
+		if bitnr in self.digPulses:
+			logger.warning('pulse on %d still running while a new one requested')
+		else:
+			self.digPulses[bitnr] = {'start':None, 'dur':duration}
+			logger.info('digital pulse on %s duration %f' % (bitnr,duration))
+		
+	async def pulseHandler(self):
+		logger.info('running aios pulse handler on %s' % self.digPulses)
+		while True:
+			if self.digPulses:
+				plsdone=[]
+				for bitnr,pls in self.digPulses.items():
+					if pls['start']:
+						if time.perf_counter()>pls['start']+pls['dur']:
+							self.setDigBit(bitnr, False)
+							plsdone.append(bitnr)
+						else:  # doing pulse high
+							await asyncio.sleep(0.02)
+					else:  
+						pls['start']=time.perf_counter()
+						logger.debug('starting pulse:%s' % pls)
+						self.setDigBit(bitnr, True)
+						await asyncio.sleep(0.02)
+				if plsdone:
+					logger.debug('dequeu pulses :%s' % plsdone)
+					for bitnr in plsdone:
+						del self.digPulses[bitnr]
+			else:
+				await asyncio.sleep(0.2)
+
 
 	async def receiveCharValue(self):
 		""" consume received notification data """
@@ -247,29 +310,40 @@ class aiosDelegate(bluepyDelegate):
 		else:
 			return chId,val
 
+async def blinkTask(aios, bitnr=5):
+	""" example task blinking led connected to pin 0.04 (A2) """
+	logger.info('running blink task')
+	while True:
+		aios.setDigPulse(bitnr, 0.9)
+		#aios.setDigBit(bitnr, True)
+		#await asyncio.sleep(0.1)
+		#aios.setDigBit(bitnr, False)
+		await asyncio.sleep(9.1)
+
 async def main():
+	DIGOUTBIT = 5  # MOSFET flash A3
+	REDLED =17
 	logger.info("Connecting...")
 	aios = aiosDelegate(DEVADDRESS)
-	logger.info("getting notified in %s" % aios)
+	
 	if aios.dev:
-		aios.startServiceNotifyers(aios.dev.getServiceByUUID(btle.UUID(ENV_SVR))) # activate environamental service
+		# activate environamental service
+		aios.startServiceNotifyers(aios.dev.getServiceByUUID(btle.UUID(ENV_SVR))) 
 		aios.setAnaVoltRange(1, 1.2)
 		aios.startChIdNotifyer(chDIGI)
 		aios.startChIdNotifyer(chANA1ST+1)  # activate A1 analog channel
+		aios.startChIdNotifyer(chTEMP)
 	try:
-		await asyncio.gather( * aios.tasks() )
+		tasks = aios.tasks()
+		tasks.append(asyncio.create_task( blinkTask(aios, DIGOUTBIT) ))
+		logger.info("running tasks %s \nin aios %s" % (tasks,aios))
+		await asyncio.gather( * tasks )
 	except KeyboardInterrupt:
 		logger.warning('leaving')
 	finally:
 		aios.dev = None
 
 if __name__ == "__main__":	# testing 
-	#aiosID = btle.UUID(AIOS_SVR)
-	#aios = dev.getServiceByUUID(aiosID)
-	#showChars(aios)
-
-	asyncio.run(main())
-
-	#dev.disconnect()
+	asyncio.run(main()) #, debug=True)
 	time.sleep(1)
 	logger.warning('bye')