+++ fsmain.py
@@ -14,6 +14,7 @@ from accessories.hue.hueHap import add_HUE_to_bridge
 from accessories.p1smart.p1DSMR_Hap import add_p1DSMR_to_bridge
 from accessories.fs20.fs20_hap import add_fs20_to_bridge
 from accessories.BLEAIOS.aiosHap import add_AIOS_to_bridge
+from accessories.netgear.netgear_HAP import add_WNDR_to_bridge
 from lib.tls import get_logger
 
 __maintainer__ = "Henk Jan van Aalderen"
@@ -30,6 +31,7 @@ add_fs20_to_bridge(bridge, config="fs20.json")
 add_p1DSMR_to_bridge(bridge, config="p1DSMR.json")
 add_HUE_to_bridge(bridge, config="hue.json")
 add_AIOS_to_bridge(bridge, config="aios.json")
+add_WNDR_to_bridge(bridge, config="WNDR.json")
 
 driver.add_accessory(accessory=bridge)